Rename view directories and make them lowercase
