+++ src/views/alert/AlertView.tsx
